code_generator/requirements.txt
psycopg2-binary == 2.9.5
pydantic == 1.10.5
pyqt6 == 6.4.2
flask == 2.2.3
celery == 5.2.7
python-logstash == 0.4.8